# Optimize performance for extension method "ToMonitoredItemNotifications"