… components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Protocol/tests/TestData/DataChangeNotification.json
… components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Protocol/tests/TestData/SubscriptionMonitoredItems.json