**Issue:**
Running the OPC Publisher for 10min, 36% of total time within extension method `ToMonitoredItemNotifications` based on multiple enumerating the IEnumerables.

**Solution:**
Create temporary dictionary for faster lookup ClientHandles.

**General:**
* Added functional test
* Added performance test for old and new implementation using benchmark.net NuGet Package- not for daily running

**Performance:**

Method |           Job | BuildConfiguration | N|        Mean |     Error |      StdDev |      Median |
--------- |-------------- |------------------- |------------------ |------------:|----------:|------------:|------------:|
Original | .NET Core 3.1 |            Default |              1000 | `14,254.4 μs` | 275.99 μs |   283.42 μs | 14,155.4 μs |
Original | .NET Core 3.1 |            Default |              1000 | `15,818.0 μs` | 368.12 μs | 1,085.40 μs | 15,951.6 μs |
New | .NET Core 3.1 |            Default |              1000 |    **153.0 μs** |   0.79 μs |     0.70 μs |    153.2 μs |
New | .NET Core 3.1 |            Default |              1000 |    **163.6 μs** |   5.58 μs |    16.45 μs |    152.6 μs |